Fixed kill

+++ scripts/docker-redeploy.sh
@@ -5,7 +5,10 @@ cd ..
 make build test push
 
 OLD_CONTAINER_ID=`docker ps | grep leozilla/indoorcup-web | awk '{print $1}'`
-docker kill $OLD_CONTAINER_ID
+if [ ! -z "$OLD_CONTAINER_ID" ];
+then
+    docker kill $OLD_CONTAINER_ID
+fi
 
 docker run -d -p 80:80 leozilla/indoorcup-web
